OSIS-607 Refactoring admission view
